% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.prep.R
\name{data.prep}
\alias{data.prep}
\title{This function prepares the data that can be used by other functions}
\usage{
data.prep(scored.data, subtest.infor)
}
\arguments{
\item{scored.data}{A dataset that contains scored responses only.}

\item{subtest.infor}{A numerical vector. The first number indicates the number of subtests,
followed by the number o items of each subtest.}
}
\value{
A list that contains datasets of all subtests and the whole test.
}
\description{
This function generates a list of datasets based on the scored original dataset,
which can be used as objects in subscore computing functions.
}
\examples{
subtest.infor<-c(3,15,15,20)
        # This test consists of 3 subtests, which have 15, 15 and 20 items respectively.
        test.datalist<-data.prep(scored.data,subtest.infor)
}

