% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegOnTotSub.R
\name{RegOnTotSub}
\alias{RegOnTotSub}
\title{Computing subscores using Haberman's method based on both observed total scores and observed subscores.}
\usage{
RegOnTotSub(test.data)
}
\arguments{
\item{test.data}{A list that contains subscale responses and the total test responses. It 
can be obtained using the function 'data.prep'.}
}
\value{
\item{summary}{Summary of obtained subscores (e.g., mean, sd).}
\item{PRMSE}{PRMSEs of obtained subscores (for Haberman's methods only).}
\item{subscore.original}{Original observed subscores and total score.} 
\item{subscore.RegOnTotSub}{Subscores that are estimated based on both the observed total score and observed subscore.}
}
\description{
This function estimate true subscores based on both observed stotal scores and observed 
subscores using the method introduced by Haberman (2008).
}
\examples{
       test.data<-data.prep(scored.data,c(3,15,15,20))
       
       RegOnTotSub(test.data) 
       
       RegOnTotSub(test.data)$summary
       RegOnTotSub(test.data)$PRMSE
       RegOnTotSub(test.data)$subscore.RegOnTotSub
}
\references{
{
Haberman, S. J. (2008). 
"When can subscores have value?."
 Journal of Educational and Behavioral Statistics, 33(2), 204-229. 
}
}

