\name{scored.data}
\alias{scored.data}
\docType{data}
\title{
Sample scored data
}
\description{
This dataset contains responses of 150 examinees to three subscales (algebra, geometry, and measurement). These subscales consist of 15, 15, and 20 items respectively.
}
\usage{data("scored.data")}
\format{
  A data frame with 150 observations on the following 50 variables.
  \describe{
    \item{\code{V1}}{Item 1}
    \item{\code{V2}}{Item 2}
    \item{\code{V3}}{Item 3}
    \item{\code{V4}}{Item 4}
    \item{\code{V5}}{Item 5}
    \item{\code{V6}}{Item 6}
    \item{\code{V7}}{Item 7}
    \item{\code{V8}}{Item 8}
    \item{\code{V9}}{Item 9}
    \item{\code{V10}}{Item 10}
    \item{\code{V11}}{Item 11}
    \item{\code{V12}}{Item 12}
    \item{\code{V13}}{Item 13}
    \item{\code{V14}}{Item 14}
    \item{\code{V15}}{Item 15}
    \item{\code{V16}}{Item 16}
    \item{\code{V17}}{Item 17}
    \item{\code{V18}}{Item 18}
    \item{\code{V19}}{Item 19}
    \item{\code{V20}}{Item 20}
    \item{\code{V21}}{Item 21}
    \item{\code{V22}}{Item 22}
    \item{\code{V23}}{Item 23}
    \item{\code{V24}}{Item 24}
    \item{\code{V25}}{Item 25}
    \item{\code{V26}}{Item 26}
    \item{\code{V27}}{Item 27}
    \item{\code{V28}}{Item 28}
    \item{\code{V29}}{Item 29}
    \item{\code{V30}}{Item 30}
    \item{\code{V31}}{Item 31}
    \item{\code{V32}}{Item 32}
    \item{\code{V33}}{Item 33}
    \item{\code{V34}}{Item 34}
    \item{\code{V35}}{Item 35}
    \item{\code{V36}}{Item 36}
    \item{\code{V37}}{Item 37}
    \item{\code{V38}}{Item 38}
    \item{\code{V39}}{Item 39}
    \item{\code{V40}}{Item 40}
    \item{\code{V41}}{Item 41}
    \item{\code{V42}}{Item 42}
    \item{\code{V43}}{Item 43}
    \item{\code{V44}}{Item 44}
    \item{\code{V45}}{Item 45}
    \item{\code{V46}}{Item 46}
    \item{\code{V47}}{Item 47}
    \item{\code{V48}}{Item 48}
    \item{\code{V49}}{Item 49}
    \item{\code{V50}}{Item 50}  }
}
\details{
A dataset containing responses of 150 examinees to a total number of 50 math items on three subscales: 15 algebra items, 15 geometry, and 20 measurement items.
}
\examples{
#read data
data(scored.data)
}
\keyword{datasets}
