\name{TIMSS11G8M.data}
\alias{TIMSS11G8M.data}
\docType{data}
\title{
The 2011 TIMSS Grade 8 Mathematics Assessment Dataset
}
\description{
The TIMSS dataset used in Dai, Svetina, and Wang (2017) (doi:10.3102/1076998617716462).It contained responses from
765 students to 32 items with 6 to 9 items on each of the subscales of (1) number (Q1 to Q9), (2) algebra (Q10 to Q18), 
(3) geometry (Q19 to Q24), and (4) data and chance (Q25 to Q30). Omitted responses were treated as incorrect.  
}
\usage{data("TIMSS11G8M.data")}
\format{
  A data frame with 765 observations on the following 32 variables.
  \describe{
    \item{\code{Q1}}{a numeric vector}
    \item{\code{Q2}}{a numeric vector}
    \item{\code{Q3}}{a numeric vector}
    \item{\code{Q4}}{a numeric vector}
    \item{\code{Q5}}{a numeric vector}
    \item{\code{Q6}}{a numeric vector}
    \item{\code{Q7}}{a numeric vector}
    \item{\code{Q8}}{a numeric vector}
    \item{\code{Q9}}{a numeric vector}
    \item{\code{Q10}}{a numeric vector}
    \item{\code{Q11}}{a numeric vector}
    \item{\code{Q12}}{a numeric vector}
    \item{\code{Q13}}{a numeric vector}
    \item{\code{Q14}}{a numeric vector}
    \item{\code{Q15}}{a numeric vector}
    \item{\code{Q16}}{a numeric vector}
    \item{\code{Q17}}{a numeric vector}
    \item{\code{Q18}}{a numeric vector}
    \item{\code{Q19}}{a numeric vector}
    \item{\code{Q20}}{a numeric vector}
    \item{\code{Q21}}{a numeric vector}
    \item{\code{Q22}}{a numeric vector}
    \item{\code{Q23}}{a numeric vector}
    \item{\code{Q24}}{a numeric vector}
    \item{\code{Q25}}{a numeric vector}
    \item{\code{Q26}}{a numeric vector}
    \item{\code{Q27}}{a numeric vector}
    \item{\code{Q28}}{a numeric vector}
    \item{\code{Q29}}{a numeric vector}
    \item{\code{Q30}}{a numeric vector}
    \item{\code{Q31}}{a numeric vector}
    \item{\code{Q32}}{a numeric vector}
  }
}
\source{
Dai, S., Svetina, D., & Wang, X. (2017). 
"Reporting subscores using R: A software review."
 Journal of Educational and Behavioral Statistics. 42(2), 617-638. doi: 10.3102/1076998617716462. 
}
\examples{
data(TIMSS11G8M.data)
}
\keyword{datasets}
