% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep.R
\name{data.prep}
\alias{data.prep}
\title{This function prepares data into a requried list format}
\usage{
data.prep(scored.data, subtest.infor)
}
\arguments{
\item{scored.data}{Original scored dataset with rows as individuals and columns as items.}

\item{subtest.infor}{A numerical vector. The first number indicates the number of subtests,
followed by numbers o items on each subscale.}
}
\value{
A list that contains subscale responses and the total test response.
}
\description{
This function generates a list of datasets using the scored original dataset, 
which can be used as objects in subscore computing functions.
}
\examples{
        subtest.infor<-c(3,15,15,20) 
        # This test consists of 3 subtests, which have 15, 15 and 20 items respectively.
        test.data<-data.prep(scored.data,subtest.infor)
}

