\name{test.data}
\alias{test.data}
\docType{data}
\title{
A list of objects that include both test information and subscores.
}
\description{
This list consists of four objects. The first three objects are item responses on the three subscales (algebra, geometry, and measurement). The fourth object is the response data on the total test.
}
\usage{data("test.data")}
\format{
  The format is:
List of 4
 $ Algebra    :'data.frame':	150 obs. of  15 variables.
 
 $ Geometry   :'data.frame':	150 obs. of  15 variables.
  
 $ Measurement:'data.frame':	150 obs. of  20 variables.
 
 $ Math       :'data.frame':	150 obs. of  50 variables.
 
 }
\details{
Algebra : Responses of 150 participants to 15 items; Geometry: Responses of 150 participants to 15 items.
Measurement: Responses of 150 participants to 20 items; Math: Responses of 150 participants to 20 items.
}
\keyword{datasets}
